+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0 (Jul 19th 2025)
+
+- decodeAccess() function now handles access token from bearer token in authorization header
+- Update Passken.js library to 0.4.0
+
 # 0.1.0 (Mar 04th 2025)
 
 - Initial release